Annex 3 PID Rulebook: Use not-equals for other than age_over_NN
On two occasions `age_over_NN` is defined with a value for `NN` other
than 18. First using "<" (less than), later using "<>", which can either
mean less or greater than or not equal to. Since there's no symbol in
between the brackets, like a / or | to indicate alternate options, I've
interpreted this to have the latter meaning.

For clarity both occurences are replaced with `≠`. This is a unicode
symbol outside of the basic ASCII character set and can cause trouble in
LaTeX documents but the modern xelatex setup in use handles it
transparently.