Annex 3 PID Rulebook: Rephrase to change meaning
The `age_over_NN` attributes should not change at every one of the
user's birthdays. That would imply they contain the exact current age,
which is more information than the attribute is intended to convey.
Rather the value of the attributes *may* change, for example when a user
turns 18 `age_over_18` would change from false to true, but when they
turn 17 or 19 there is no change.

While `age_in_years` does change on every birthday, stating that it may
change on every birthday is still correct.