Fix formatting of continued list
The continued list in section 6.6.3.1 was not being rendered as a list but as a
continuation of the preceding paragraph. Adding the blank line makes it render
as a list, but instead breaks the continued item numbering in MkDocs. Pandoc
respects the starting number of the continued list, but MkDocs does not by
default. This therefore requires the [`sane_lists`][1] MkDocs extension.

Enabling the `sane_lists` extension could have unintended consequences, so I
compared the "main" ARF page of the rendered MkDocs sites side by side to check.
I found only one place affected apart from the intended one in 6.6.3.1, namely a
nested bullet point list in section 6.6.3.9 which was being rendered as numbered
items of the parent list. This list formatting was broken by enabling the
`sane_lists` extension, but fixed by indenting the nested lists to 4 spaces
instead of 3.

[1]: https://python-markdown.github.io/extensions/sane_lists/