Release/2.5.0 (#602)
* Release/2.5.0 (#1454)

* Updated the links (#1444)

* Feature/2.5.0 (#1453)

* Fixing broken reference

* Solving #1342

* Correction typo in references to Topic 31

* Solving #1396

* Update architecture-and-reference-framework-main.md

Fixing description of the use of registration certificates in section 6.6.3.3 to make clear that an RP can have multiple reg certs and that, if there is no reg cert, the Wallet Unit needs to have the intended use to be able to request the registered attributes from the Registrar.

* Improving the wording of RPRC_09 and RPRC_13

* Solving #1406  and #1341

For #1406: Changes in DASH_03, RPT_DPA_05 and _05a, and RPRC_12

For  #1341: Changes to Topic 27 introduction (not finished yet)

* Solved #1341

* Final changes for #1341

* Solved #1399

* Solving #1352

* Solving #1292

* Solving #1424

* Removing "(sub-)entitlements" from ARF

Harri noted that the words "(sub-)entitlement" are not used in TS5 and TS6, and they are used in a different meaning in ETSI 119 475 and 119 495. Therefore, I have removed these terms from the ARF, and used sligthly more verbose descriptions instead.

* Changes regarding URL of RP's Registrar in case an intermediary is involved

Harri pointed out that DASH_03 point f) is not correct in case an intermediary is involved. I have added a note to clarify how the Wallet Unit can obtain this URL in case an intermediary is involved. I added similar notes to RPRC_18 and RPRC_19. I added RPRC_08a to ensure a registration cert contains this URL.

Also, some more minor fixes as a result of Harri's review.

* Solving #1283

* Solving #1156

* Solving #1087

* Solved #1282

* Solving #1387

* Solved #1288

* Solved #1391

* Adding definition of 'intermediary' to Annex 1

See also #1159

* Editorial (markdown) corrections

First changes for fixing #1245 in README file

* Fixing #1245

* Correcting description of how OpenID4VP uses attestation types

* Solving #1284

* Solving #1253

* Accessibility update to chapter 8

main doc, annex2, annex3.1, annex3.2

* Update architecture-and-reference-framework-main.md

Typo in link

* Update architecture-and-reference-framework-main.md

Update on change log

* Topic o and z discussion papers updated, release 3 (#597)

* Updated topic O discussion paper r3

* Updated topic Z discussion paper r3

* age verfication attributes removed solving ##1429 Revision table added (#1442)

Co-authored-by: Nikos <nikosft@gmail.com>

* Changelog

* Update references

* rulebooks moved on a dedicated repository

* polishing and linting

* polishing and linting

* polishing and linting

* polishing and linting

* polishing and linting

---------

Co-authored-by: David Bakker <75264147+david-bakker@users.noreply.github.com>
Co-authored-by: Paul Hin <phin@scytales.com>
Co-authored-by: Nikos <nikosft@gmail.com>

---------

Co-authored-by: Paul Hin <phin@scytales.com>
Co-authored-by: David Bakker <75264147+david-bakker@users.noreply.github.com>
Co-authored-by: Nikos <nikosft@gmail.com>

* Links fixed (#1455)

* fixing links and configuration

* fixing links, linting and cleaning

* added media annex 5

* fixing links, linting and cleaning

---------

Co-authored-by: Paolo De Rosa <paolo.de.rosa@linux.com>
Co-authored-by: David Bakker <75264147+david-bakker@users.noreply.github.com>
Co-authored-by: Nikos <nikosft@gmail.com>